requirements.txt
pypresence>=4.3.0
winsdk>=1.0.0b10
winshell>=0.6
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
pyinstaller>=6.0.0
